refactor: move use-data-value-params
